Reorganisation of simulation v2 files
Spring cleaning!
