+++ packages/simulation-v2/src/FSAStep.ts
@@ -1,0 +1,10 @@
+import { FSAGraphProblem } from "./FSAGraph";
+import { UnparsedFSAGraph } from "./graph";
+import { parseGraph } from "./parse-graph";
+
+// export const graphStepper = (graph: UnparsedFSAGraph, input: string) => {
+//     const parsedGraph = parseGraph(graph);
+//     const problem = new FSAGraphProblem(parsedGraph, input);
+//     const stepper = new graphStepper(problem);
+//     return stepper;
+// };
